requirements.txt
phidata
agno
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
streamlit
faiss-cpu
sentence-transformers
psycopg2-binary
pandas
PyPDF2
betterstack